+++ src/completion.ts
@@ -1,6 +1,10 @@
+import {commonHtmlAttrs, htmlAttrs, extAttrs} from 'wikiparser-node/util/sharable.ts';
 import type * as Monaco from 'monaco-editor';
 import type {languages, Position} from 'monaco-editor';
 import type {Config} from 'wikiparser-node';
+
+// eslint-disable-next-line @stylistic/max-len
+const re = /(?:<\/?(\w+)|\{\{\s*(#[^|{}<>[\]#]*)|(__(?:(?!__)[\p{L}\d_])+)|(?:^|[^[])\[([a-z:/]+)|<(\w+)(?:\s(?:[^<>{}|=\s]+(?:\s*=\s*(?:[^\s"']\S*|(["']).*?\6))?(?=\s))*)?\s(\w+))$/iu;
 
 const completion = (monaco: typeof Monaco): languages.CompletionItemProvider => {
 	monaco.editor.addKeybindingRule({
@@ -18,12 +22,12 @@ const completion = (monaco: typeof Monaco): languages.CompletionItemProvider => 
 	 * @param pos.column 列号
 	 */
 	const getCompletion = (
-		words: string[],
+		words: Iterable<string>,
 		kind: keyof typeof languages.CompletionItemKind,
 		mt: string,
 		{lineNumber, column}: Position,
 	): languages.CompletionList => ({
-		suggestions: words.map(w => ({
+		suggestions: [...new Set(words)].map(w => ({
 			label: w,
 			kind: monaco.languages.CompletionItemKind[kind],
 			insertText: w,
@@ -32,7 +36,8 @@ const completion = (monaco: typeof Monaco): languages.CompletionItemProvider => 
 	});
 
 	let config: Config | undefined,
-		tags: string[],
+		tags: Set<string>,
+		allTags: string[],
 		functions: string[],
 		switches: string[],
 		protocols: string[];
@@ -46,24 +51,48 @@ const completion = (monaco: typeof Monaco): languages.CompletionItemProvider => 
 			} else if (!config?.namespaces[1]) {
 				config = await wikiparse.getConfig(); // eslint-disable-line require-atomic-updates
 				const {ext, html, parserFunction, doubleUnderscore, protocol} = config;
-				tags = [ext, html, 'onlyinclude', 'includeonly', 'noinclude'].flat(2);
+				tags = new Set([ext, html].flat(2));
+				allTags = [...tags, 'onlyinclude', 'includeonly', 'noinclude'];
 				functions = [Object.keys(parserFunction[0]), parserFunction.slice(1) as string[][]].flat(2);
 				switches = (doubleUnderscore.slice(0, 2) as string[][]).flat().map(w => `__${w}__`);
 				protocols = protocol.split('|');
 			}
 			const before = model.getValueInRange(new monaco.Range(pos.lineNumber, 1, pos.lineNumber, pos.column)),
-				re = /(?:<\/?(\w+)|\{\{\s*(#[^|{}<>[\]#]*)|(__(?:(?!__)[\p{L}\d_])+)|(?:^|[^[])\[([a-z:/]+))$/iu,
 				mt = re.exec(before);
 			if (!mt) {
 				return null;
 			} else if (mt[1]) { // tag
-				return getCompletion(tags, 'Class', mt[1], pos);
+				return getCompletion(allTags, 'Class', mt[1], pos);
 			} else if (mt[2]) { // parser function
 				return getCompletion(functions, 'Function', mt[2], pos);
 			} else if (mt[3]) { // behavior switch
 				return getCompletion(switches, 'Constant', mt[3], pos);
 			} else if (mt[4]) { // protocol
 				return getCompletion(protocols, 'Reference', mt[4], pos);
+			} else if (mt[5]) { // attribute key
+				const tag = mt[5].toLowerCase();
+				if (!tags.has(tag)) {
+					return null;
+				}
+				const key = mt[7]!,
+					thisHtmlAttrs = htmlAttrs[tag],
+					thisExtAttrs = extAttrs[tag],
+					extCompletion = thisExtAttrs ? getCompletion(thisExtAttrs, 'Field', key, pos) : null;
+				return config.ext.includes(tag) && !thisHtmlAttrs
+					? extCompletion
+					: {
+						suggestions: [
+							...extCompletion?.suggestions ?? [],
+							...tag === 'meta' || tag === 'link'
+								? []
+								: getCompletion(commonHtmlAttrs, 'Property', key, pos).suggestions,
+							...thisHtmlAttrs
+								? getCompletion(thisHtmlAttrs, 'Property', key, pos).suggestions
+								: [],
+							...getCompletion(['data-'], 'Variable', key, pos).suggestions,
+							...getCompletion(['xmlns:'], 'Interface', key, pos).suggestions,
+						],
+					};
 			}
 			throw new Error('Unknown completion type!');
 		},