+++ src/tree.ts
@@ -6,10 +6,10 @@ declare type Tree = Promise<AST> & {docChanged?: boolean};
 
 const trees = new WeakMap<editor.ITextModel, Tree>();
 
-export const getTree = (model: editor.ITextModel, stage = 8): Tree => {
+export const getTree = (model: editor.ITextModel): Tree => {
 	let tree = trees.get(model);
 	if (!tree || tree.docChanged) {
-		tree = wikiparse.json(model.getValue(), true, -6, stage);
+		tree = wikiparse.json(model.getValue(), true, -6, 8);
 		trees.set(model, tree);
 	}
 	return tree;